forgot to add ascii implementation back lol
