+++ bot.py
@@ -3,6 +3,9 @@ from discord.ext import commands
 import io
 import os
 import logging
+from PIL import Image
+
+from scripts import ascii
 
 print("Brewiiz Bot Runner\n (C) 2025 Brewy. Some Rights Reserved.")
 
@@ -13,6 +16,37 @@ intents.message_content = True
 allowed_mentions = discord.AllowedMentions(everyone = True)
 bot = commands.Bot(command_prefix=';', intents=intents)
 
+@bot.command()
+async def ascii_gen(ctx):
+    await ctx.send("Initializing :3")
+    if ctx.message.attachments:
+        attachment = ctx.message.attachments[0]
+        try:
+            image_bytes = await attachment.read()
+            image = Image.open(io.BytesIO(image_bytes))
+
+            if image.format not in ['JPEG', 'PNG']:
+                await ctx.send('Unsupported image format. Please upload a JPEG or PNG image.')
+                return
+
+         
+            await ctx.send("we're doing ts... one secondish :]")
+
+       
+            ascii_art = ascii.image_to_ascii(image, ascii.ascii_chars_light_to_dark, max_width=40)
+
+            
+            max_embed_length = 4096 - 10 
+            if len(ascii_art) > max_embed_length:
+                ascii_art = ascii_art[:max_embed_length]  
+            embed = discord.Embed(title="Result", description=f'```\n{ascii_art}\n```')
+            
+            await ctx.send(embed=embed)
+            await ctx.send("generated °3°")
+        except Exception as e:
+            await ctx.send(f'# Oops. \nan error occurred. Details:\n{str(e)}')
+    else:
+        await ctx.send(" attach an image. i can't generate anything mate :(")
 
 @bot.event
 async def on_ready():
